feat: split hyperopt class
this ensures it's clear which parts are passed to workers
